# docs/swagger.yaml
basePath: /
definitions:
  accounting.TransferItem:
    properties:
      amount:
        type: integer
      created_at:
        type: string
      direction:
        type: string
      from_bank_account:
        type: string
      id:
        type: integer
      to_bank_account:
        type: string
    type: object
  accounting.TransferRequest:
    properties:
      amount:
        type: integer
      bank_account:
        type: string
    type: object
  user.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  user.MeResponse:
    properties:
      bank_account:
        type: string
      credit:
        type: integer
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      username:
        type: string
    type: object
  user.RegisterRequest:
    properties:
      bank_account:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  user.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      token_type:
        type: string
    type: object
  user.UpdateRequest:
    properties:
      bank_account:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
    type: object
  user.User:
    properties:
      bank_account:
        type: string
      created_at:
        type: string
      credit:
        type: integer
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  utils.ErrorResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:5000
info:
  contact: {}
  description: Money Transfer API
  title: ProjectTest API
  version: "1.0"
paths:
  /accounting/transfer:
    post:
      consumes:
      - application/json
      description: Transfer money to another user
      parameters:
      - description: Transfer request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/accounting.TransferRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              format: int64
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - Bearer: []
      summary: Create transfer
      tags:
      - Accounting
  /accounting/transfer-list:
    get:
      consumes:
      - application/json
      description: Get list of transfers for the current user
      parameters:
      - description: Start date
        in: query
        name: start_date
        type: string
      - description: End date
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/accounting.TransferItem'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - Bearer: []
      summary: List transfers
      tags:
      - Accounting
  /user/login:
    post:
      consumes:
      - application/json
      description: User login with username and password
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Login user
      tags:
      - User
  /user/me:
    get:
      consumes:
      - application/json
      description: Get information about the logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.MeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - Bearer: []
      summary: Get current user info
      tags:
      - User
  /user/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Register user
      tags:
      - User
  /user/update:
    patch:
      consumes:
      - application/json
      description: Update user profile information
      parameters:
      - description: Update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - Bearer: []
      summary: Update user info
      tags:
      - User
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"